Split hooks into separate files and add their respective types;
