+++ components/jrg/auth/hooks/useUser.ts
@@ -1,33 +1,51 @@
-'use client';
-import axios from 'axios';
-import { getCookie, setCookie } from 'cookies-next';
-import useSWR from 'swr';
-import { useRouter } from 'next/navigation';
+import { createGraphQLClient } from '@/components/interactive/hooks';
+import { getCookie } from 'cookies-next';
+import useSWR, { SWRResponse } from 'swr';
+import log from '../../next-log/log';
+import { User, UserSchema } from '../types/types';
 
-// Create a custom SWR hook for a specific endpoint
-export default function useUser() {
-  const router = useRouter();
-  return useSWR(
-    '/user',
-    async () => {
-      const userDataReq = await axios.get(`${process.env.NEXT_PUBLIC_AGIXT_SERVER}/v1/user`, {
-        headers: {
-          Authorization: getCookie('jwt'),
-        },
-      });
-      // if (!getCookie('agixt-company-id')) {
-      //   console.log('Setting company ID to: ' + userDataReq.data.companies.filter((x) => x.primary)[0].id);
-      //   setCookie('agixt-company-id', userDataReq.data.companies.filter((x) => x.primary)[0].id, {
-      //     domain: process.env.NEXT_PUBLIC_COOKIE_DOMAIN,
-      //   });
-      // }
-      if (userDataReq.status === 402) {
-        router.push('/user/subscribe');
+/**
+ * Hook to fetch and manage current user data
+ * @returns SWR response containing user data
+ */
+export function useUser(): SWRResponse<User | null> {
+  const client = createGraphQLClient();
+
+  return useSWR<User | null>(
+    ['/user', getCookie('jwt')],
+    async (): Promise<User | null> => {
+      if (!getCookie('jwt')) return null;
+      try {
+        const query = UserSchema.toGQL('query', 'GetUser');
+        log(['GQL useUser() Query', query], {
+          client: 3,
+        });
+        const response = await client.request<{ user: User }>(query);
+        log(['GQL useUser() Response', response], {
+          client: 3,
+        });
+        return UserSchema.parse(response.user);
+      } catch (error) {
+        log(['GQL useUser() Error', error], {
+          client: 1,
+        });
+        return {
+          companies: [],
+          email: '',
+          firstName: '',
+          id: '',
+          lastName: '',
+        };
       }
-      return userDataReq.data;
     },
     {
-      fallbackData: {},
+      fallbackData: {
+        companies: [],
+        email: '',
+        firstName: '',
+        id: '',
+        lastName: '',
+      },
     },
   );
 }
